==> Codigos/9. XGBoost.R <==
#------------------------------------------------------------------------------#
#----------------------------- Pruebas - XG Boosting  -------------------------#
#------------------------------------------------------------------------------#

# 0. Carga de informacion  -----------------------------------------------------
setwd(paste0(wd,"/Base"))
Train <- import(file = "Train_hogares_final.rds")
Test <- import(file = "Test_hogares_final.rds")

# 1. Carga de informaci?n balanceada (SMOTE) -----------------------------------
# Este procedimiento se realizo en el c?digo 2.Pruebas_Elastic_Net
setwd(paste0(wd,"\\Base\\Base_Elastic_Net"))
Train_3_SMOTE = import(file = "Train_3_SMOTE.rds")

# Cargar test_2
set.seed(1536) # Para reproducibilidad
Train_3_indices <- as.integer(createDataPartition(Train$Pobre, p = 0.85, list = FALSE))
Train_3 <- Train[Train_3_indices, ]
Test_3 <- Train[-Train_3_indices, ]
prop.table(table(Train_3$Pobre))
prop.table(table(Test_3$Pobre))

# Convertir las variables categ?ricas a factores en Test
Dummys <- subset(Test, select = c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                  Head_Oficio,Head_Ocupacion))

Dummys$Dominio <- as.factor(Dummys$Dominio)
Dummys$tipo_vivienda <- as.factor(Dummys$tipo_vivienda)
Dummys$maxEducLevel <- as.factor(Dummys$maxEducLevel)
Dummys$Head_EducLevel <- as.factor(Dummys$Head_EducLevel)
Dummys$Head_Oficio <- as.factor(Dummys$Head_Oficio)
Dummys$Head_Ocupacion <- as.factor(Dummys$Head_Ocupacion)

Dummys <- model.matrix(~ . - 1, data = Dummys)  # Eliminar el intercepto

# Union con la base original
Test <- cbind(subset(Test, select = -c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                       Head_Oficio,Head_Ocupacion)),Dummys)

# Convertir las variables categ?ricas a factores en test_3
Dummys <- subset(Test_3, select = c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                    Head_Oficio,Head_Ocupacion))

Dummys$Dominio <- as.factor(Dummys$Dominio)
Dummys$tipo_vivienda <- as.factor(Dummys$tipo_vivienda)
Dummys$maxEducLevel <- as.factor(Dummys$maxEducLevel)
Dummys$Head_EducLevel <- as.factor(Dummys$Head_EducLevel)
Dummys$Head_Oficio <- as.factor(Dummys$Head_Oficio)
Dummys$Head_Ocupacion <- as.factor(Dummys$Head_Ocupacion)

Dummys <- model.matrix(~ . - 1, data = Dummys)  # Eliminar el intercepto

# Union con la base original
Test_3 <- cbind(subset(Test_3, select = -c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                           Head_Oficio,Head_Ocupacion)),Dummys)


# 2. Se paraleliza -------------------------------------------------------------
library(doParallel)
num_cores <- parallel::detectCores()
print(num_cores)
cl <- makeCluster(3)  # Usar todos menos 3
registerDoParallel(cl)
#stopCluster(cl)    # Para liberar los nucleos

#3.Estimaciones ----------------------------------------------------------------
#3.1.1 Estimacion 1 ------------------------------------------------------------

fiveStats <- function(...)  c(defaultSummary(...),  prSummary(...)) 

fitControl<- trainControl(method = "cv",
                          number = 5,
                          classProbs = TRUE,
                          summaryFunction = fiveStats,
                          savePredictions = "final")

grid_xbgoost <- expand.grid(nrounds = c(100,500),
                            max_depth = c(2,4,6), 
                            eta = c(0.01,0.05,0.1), 
                            gamma = c(0,0.5,1), 
                            min_child_weight = c(10,20,30),
                            colsample_bytree = c(0.33,0.66),
                            subsample = c(0.4,0.8))

#3.1.2 Modelo 1 ----------------------------------------------------------------
Regresores_1 = c(colnames(Train_3_SMOTE)[1:26],colnames(Train_3_SMOTE)[35:43],colnames(Train_3_SMOTE)[69:85],
                 colnames(Train_3_SMOTE)[168:176])

set.seed(1536)
Xgboost_tree_1 <- train(formula(paste0("class ~", paste0(Regresores_1, collapse = " + "))),
                        data=Train_3_SMOTE,
                        method = "xgbTree", 
                        trControl = fitControl,
                        tuneGrid=grid_xbgoost,
                        metric="F"
)        

# Mejor modelo 
Xgboost_tree_1
Xgboost_tree_1$bestTune

# Prediccion fuera de muestra
Prediccion_1_Xgboost <- predict(Xgboost_tree_1, 
                                newdata = Test_3)
# F1-Score 
f1_score_modelo_xgboost_1 <- F1_Score(y_true = as.factor(Test_3$Pobre), y_pred = as.factor(Prediccion_1_Xgboost), positive = "Yes")

# Prediccion con test 
Prediccion_1_xgboost <- predict(Xgboost_tree_1,
                                newdata = Test)

# Se deja en el formato requerido
Prediccion_1_xgboost <- as.data.frame(Prediccion_1_xgboost) %>% cbind(Test["id"]) %>%
  mutate(pobre=ifelse(Prediccion_1_xgboost =="Yes",1,0)) %>% 
  select(id,pobre)


Nombre <- paste0("Xgboost _nrounds_",500, "_maxdepth_",6,"_eta_",0.05,"_gamma_",0,
                 "_colsample_bytree_",0.33,"_min_child_weight_",10,"_subsample_",0.8,".csv") 
setwd(paste0(wd,"\\Output\\XGboost"))
write.csv(Prediccion_1_xgboost,Nombre, row.names = FALSE)

#3.2.1 Estimacion 2 ------------------------------------------------------------

fiveStats <- function(...)  c(defaultSummary(...),  prSummary(...)) 

fitControl<- trainControl(method = "cv",
                          number = 5,
                          classProbs = TRUE,
                          summaryFunction = fiveStats,
                          savePredictions = "final")

grid_xbgoost <- expand.grid(nrounds = c(300,500),
                            max_depth = c(6), 
                            eta = c(0.05), 
                            gamma = c(0,0.5), 
                            min_child_weight = c(10,20),
                            colsample_bytree = c(0.33,0.66),
                            subsample = c(0.4,0.8))

#3.2.2 Modelo 2 ----------------------------------------------------------------
Regresores_2 = c(colnames(Train_3_SMOTE)[1:26],colnames(Train_3_SMOTE)[35:43],colnames(Train_3_SMOTE)[69:85],
                 colnames(Train_3_SMOTE)[168:176],"Head_edad*Head_edad","Head_Mujer*Head_edad")

set.seed(1536)
Xgboost_tree_2 <- train(formula(paste0("class ~", paste0(Regresores_2, collapse = " + "))),
                        data=Train_3_SMOTE,
                        method = "xgbTree", 
                        trControl = fitControl,
                        tuneGrid=grid_xbgoost,
                        metric="F"
)        

# Mejor modelo 
Xgboost_tree_2
Xgboost_tree_2$bestTune

# Prediccion fuera de muestra
Prediccion_2_Xgboost <- predict(Xgboost_tree_2, 
                                newdata = Test_3)

# F1-Score 
f1_score_modelo_xgboost_2 <- F1_Score(y_true = as.factor(Test_3$Pobre), y_pred = as.factor(Prediccion_2_Xgboost), positive = "Yes")

# Prediccion con test 
Prediccion_2_xgboost <- predict(Xgboost_tree_2,
                                newdata = Test)

# Se deja en el formato requerido
Prediccion_2_xgboost <- as.data.frame(Prediccion_2_xgboost) %>% cbind(Test["id"]) %>%
  mutate(pobre=ifelse(Prediccion_2_xgboost =="Yes",1,0)) %>% 
  select(id,pobre)

Nombre <- paste0("Xgboost _nrounds_",500, "_maxdepth_",6,"_eta_",0.05,"_gamma_",0,
                 "_colsample_bytree_",0.66,"_min_child_weight_",20,"_subsample_",0.8,".csv") 
setwd(paste0(wd,"\\Output\\XGboost"))
write.csv(Prediccion_2_xgboost,Nombre, row.names = FALSE)


#4. Mejor modelo ---------------------------------------------------------------
# El mejor modelo de acuerdo con el punto de Kaggle fue el Prediccion_2_Xgboost
# Se procede a correr este modelo con los hiperpar?metros que llevaron 
# a la mejor predicci?n fuera de muestra. 

# 4. Estimacion VF -------------------------------------------------------------
Regresores_VF = c(colnames(Train_3_SMOTE)[1:26],colnames(Train_3_SMOTE)[35:43],colnames(Train_3_SMOTE)[69:85],
                 colnames(Train_3_SMOTE)[168:176],"Head_edad*Head_edad","Head_Mujer*Head_edad")

fitControl<- trainControl(method = "cv",
                          number = 5,
                          classProbs = TRUE,
                          savePredictions = "final")

grid_xgboost <- expand.grid(nrounds = 500,
                            max_depth = 6, 
                            eta = 0.05, 
                            gamma = 0, 
                            min_child_weight = 20,
                            colsample_bytree = 0.66,
                            subsample = 0.8)

Xgboost_tree_VF <- train(formula(paste0("class ~", paste0(Regresores_VF, collapse = " + "))),
                         data = Train_3_SMOTE,
                         method = "xgbTree", 
                         trControl = fitControl,
                         tuneGrid = grid_xgboost,  # Especificamos los hiperpar?metros aqu?
                         metric = "F"
)

# Prediccion dentro de muestra
Xgboost_tree_prediccion_DM <- predict(Xgboost_tree_VF, 
                                      newdata = Train_3_SMOTE)
# F1-Score 
f1_score_modelo_xgboost_DM <- F1_Score(y_true = as.factor(Train_3_SMOTE$class), y_pred = as.factor(Xgboost_tree_prediccion_DM), positive = "Yes")

# 4. Estimacion con interacciones adicionales -------------------------------------------------------------
library(doParallel)
num_cores <- parallel::detectCores()
print(num_cores)
cl <- makeCluster(6)  # Usar todos menos 3
registerDoParallel(cl)
#stopCluster(cl)    # Para liberar los nucleos
gc()

Regresores_VF = c(colnames(Train_3_SMOTE)[1:26],colnames(Train_3_SMOTE)[35:43],colnames(Train_3_SMOTE)[69:85],
                  colnames(Train_3_SMOTE)[168:176],"Head_edad*Head_edad","Head_Mujer*Head_edad", "rsubsidiado*Pago_Arriendo", 
                  "Head_Reg_subs_salud*nocupados", "Pago_Arriendo^2")

fitControl<- trainControl(method = "cv",
                          number = 5,
                          classProbs = TRUE,
                          savePredictions = "final")

grid_xgboost <- expand.grid(nrounds = c(500, 1000),
                            max_depth = c(4, 6, 8), 
                            eta = c(0.01, 0.02, 0.05), 
                            gamma = c(0, 0.1, 0.5), 
                            min_child_weight = c(10, 20),
                            colsample_bytree = c(0.5, 0.66, 0.8),
                            subsample = c(0.6, 0.7, 0.8))

Xgboost_tree_VF <- train(formula(paste0("class ~", paste0(Regresores_VF, collapse = " + "))),
                         data = Train_3_SMOTE,
                         method = "xgbTree", 
                         trControl = fitControl,
                         tuneGrid = grid_xgboost,  # Especificamos los hiperpar?metros aqu?
                         metric = "F"
)

# Prediccion dentro de muestra
Xgboost_tree_prediccion_DM <- predict(Xgboost_tree_VF, 
                                      newdata = Train_3_SMOTE)
# F1-Score 
f1_score_modelo_xgboost_DM <- F1_Score(y_true = as.factor(Train_3_SMOTE$class), y_pred = as.factor(Xgboost_tree_prediccion_DM), positive = "Yes")
f1_score_modelo_xgboost_DM

# Prediccion fuera de muestra
Xgboost_tree_prediccion <- predict(Xgboost_tree_VF, 
                           newdata = Test_3)
# F1-Score 
f1_score_modelo_xgboost_VF <- F1_Score(y_true = as.factor(Test_3$Pobre), y_pred = as.factor(Xgboost_tree_prediccion), positive = "Yes")
f1_score_modelo_xgboost_VF

# 4. Estimacion VF con nuevas variables-------------------------------------------------------------
library(doParallel)
num_cores <- parallel::detectCores()
print(num_cores)
cl <- makeCluster(6)  # Usar todos menos 3
registerDoParallel(cl)
#stopCluster(cl)    # Para liberar los nucleos
gc()

#--------------------------------

# Convertir las variables categ?ricas a factores en Test
Dummys <- subset(Test, select = c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                  Head_Oficio,Head_Ocupacion))

Dummys$Dominio <- as.factor(Dummys$Dominio)
Dummys$tipo_vivienda <- as.factor(Dummys$tipo_vivienda)
Dummys$maxEducLevel <- as.factor(Dummys$maxEducLevel)
Dummys$Head_EducLevel <- as.factor(Dummys$Head_EducLevel)
Dummys$Head_Oficio <- as.factor(Dummys$Head_Oficio)
Dummys$Head_Ocupacion <- as.factor(Dummys$Head_Ocupacion)

Dummys <- model.matrix(~ . - 1, data = Dummys)  # Eliminar el intercepto

# Union con la base original
Test <- cbind(subset(Test, select = -c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                       Head_Oficio,Head_Ocupacion)),Dummys)

# Convertir las variables categ?ricas a factores en test_3
Dummys <- subset(Test, select = c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                    Head_Oficio,Head_Ocupacion))

Dummys$Dominio <- as.factor(Dummys$Dominio)
Dummys$tipo_vivienda <- as.factor(Dummys$tipo_vivienda)
Dummys$maxEducLevel <- as.factor(Dummys$maxEducLevel)
Dummys$Head_EducLevel <- as.factor(Dummys$Head_EducLevel)
Dummys$Head_Oficio <- as.factor(Dummys$Head_Oficio)
Dummys$Head_Ocupacion <- as.factor(Dummys$Head_Ocupacion)

Dummys <- model.matrix(~ . - 1, data = Dummys)  # Eliminar el intercepto

# Union con la base original
Test <- cbind(subset(Test, select = -c(Dominio,tipo_vivienda,maxEducLevel,Head_EducLevel,
                                           Head_Oficio,Head_Ocupacion)),Dummys)


#------

Train_3_SMOTE$niños_nper <- Train_3_SMOTE$nmenores_6 / Train_3_SMOTE$Nper
Train_3_SMOTE$ocupados_nper <- Train_3_SMOTE$nocupados / Train_3_SMOTE$Nper
Train_3_SMOTE$pa_nper <- Train_3_SMOTE$nocupados / Train_3_SMOTE$Pago_Arriendo

Test_3$niños_nper <- Test_3$nmenores_6 / Test_3$Nper
Test_3$ocupados_nper <- Test_3$nocupados / Test_3$Nper
Test_3$pa_nper <- Test_3$nocupados / Test_3$Pago_Arriendo

Test$niños_nper <- Test$nmenores_6 / Test$Nper
Test$ocupados_nper <- Test$nocupados / Test$Nper
Test$pa_nper <- Test$nocupados / Test$Pago_Arriendo

# Verifica si las columnas nmenores_6 y nper existen
summary(Train_3_SMOTE$nmenores_6)
summary(Train_3_SMOTE$Nper)



Regresores_VF = c(colnames(Train_3_SMOTE)[1:50],colnames(Train_3_SMOTE)[69:85],
                  colnames(Train_3_SMOTE)[168:176], "niños_nper", "ocupados_nper",
                  "Head_edad*Head_edad","Head_Mujer*Head_edad", "Head_Afiliado_SS*Head_edad")

fitControl<- trainControl(method = "cv",
                          number = 5,
                          classProbs = TRUE,
                          savePredictions = "final")

grid_xgboost <- expand.grid(nrounds = 500,
                            max_depth = 6, 
                            eta = 0.05, 
                            gamma = 0, 
                            min_child_weight = 20,
                            colsample_bytree = 0.66,
                            subsample = 0.8)

Xgboost_tree_VF <- train(formula(paste0("class ~", paste0(Regresores_VF, collapse = " + "))),
                         data = Train_3_SMOTE,
                         method = "xgbTree", 
                         trControl = fitControl,
                         tuneGrid = grid_xgboost,  # Especificamos los hiperpar?metros aqu?
                         metric = "F"
)

# Prediccion dentro de muestra
Xgboost_tree_prediccion_DM <- predict(Xgboost_tree_VF, 
                                      newdata = Train_3_SMOTE)
# F1-Score 
f1_score_modelo_xgboost_DM <- F1_Score(y_true = as.factor(Train_3_SMOTE$class), y_pred = as.factor(Xgboost_tree_prediccion_DM), positive = "Yes")

# Prediccion fuera de muestra
Xgboost_tree_prediccion <- predict(Xgboost_tree_VF, 
                                   newdata = Test_3)
# F1-Score 
f1_score_modelo_xgboost_VF <- F1_Score(y_true = as.factor(Test_3$Pobre), y_pred = as.factor(Xgboost_tree_prediccion), positive = "Yes")
f1_score_modelo_xgboost_VF

#Prediccion de exportar

Test$DominioBOGOTA <- 0

Prediccion_ult_xgboost <- predict(Xgboost_tree_VF,
                                newdata = Test)

# Se deja en el formato requerido
Prediccion_ult_xgboost <- as.data.frame(Prediccion_ult_xgboost) %>% cbind(Test["id"]) %>%
  mutate(pobre=ifelse(Prediccion_ult_xgboost =="Yes",1,0)) %>% 
  select(id,pobre)

Nombre <- paste0("Xgboost _nrounds_",500, "_maxdepth_",8,"_eta_",0.02,"_gamma_",0,
                 "_colsample_bytree_",0.66,"_min_child_weight_",20,"_subsample_",0.8,"_ult_",".csv") 
setwd("C:/Users/Steven Ramirez/Downloads/Problem-Set-2_Machine-Learning_2024/Problem-Set-2_Machine-Learning_2024/Output/XGboost")
write.csv(Prediccion_ult_xgboost,Nombre, row.names = FALSE)


# EMBEDING ----------------------

# Predicciones de Random Forest en el conjunto de prueba
rf_predictions <- predict(RF_VF, newdata = Test_3, type = "prob")
xgb_predictions <- predict(Xgboost_tree_VF, newdata = Test_3, type = "prob")

# Predicciones de clase para ambos modelos
xgb_class_predictions <- predict(Xgboost_tree_VF, newdata = Test_3, type = "raw")
rf_class_predictions <- predict(RF_VF, newdata = Test_2, type = "raw")

# Voto mayoritario
ensemble_class_predictions <- ifelse(xgb_class_predictions == rf_class_predictions, xgb_class_predictions, 
                                     ifelse(xgb_predictions[,2] > 0.5, "Yes", "No"))

length(Test_3$Clase)
length(ensemble_class_predictions)

# F1-Score del ensemble
f1_ensemble <- F1_Score(y_true = Test_3$Pobre, y_pred = ensemble_class_predictions, positive = "Yes")

# Imprimir el F1-Score
print(f1_ensemble)

#4.1 Diagrama de algunos de los ?rboles ---------------------------------------
p_load(DiagrammeR)
tree_plot <- xgb.plot.tree(model = Xgboost_tree_VF$finalModel,
                           trees = 1, plot_width = 1000, plot_height = 500)
tree_plot

#4.2 Importancia de iteraciones ------------------------------------------------
importance <- varImp(Xgboost_tree_VF, scale = FALSE)
print(importance)
